+++ Assets/Core/SVRules/OpCodes/Type0.cs
@@ -7,7 +7,7 @@ namespace OpCode
     public class Type0 : IOpCode
     {
 
-        public int Evaluate(SVDataPacket data, List<IOpCode> operands)
+        public float Evaluate(SVDataPacket data, List<IOpCode> operands)
         {
             return Mathf.RoundToInt(data.d0.Sum(d => d.GetValue()));
         }
